feat: add RUN_CLEANUP_OPTION to run.sh script

+++ 8086sim/run.sh
@@ -1,6 +1,7 @@
 #!/bin/bash
 
 FILEPATH=$1 
+RUN_CLEANUP_OPTION=$2 
 
 BASE_FILENAME="${FILEPATH##*/}"
 BASE_FILENAME="${BASE_FILENAME%.*}"
@@ -17,6 +18,8 @@ diff "$DIR/temp" "$DIR/$BASE_FILENAME.out" && \
 echo "Done"
 
 # Do cleanup in any case
-echo "Cleanup" && \
-rm "$DIR/temp" "$DIR/$BASE_FILENAME.out" "$DIR/$BASE_FILENAME" && \
-echo "Done" 
+echo "Cleanup" 
+if [ "$RUN_CLEANUP_OPTION" = "-c" ]; then  
+    rm "$DIR/temp" "$DIR/$BASE_FILENAME.out" "$DIR/$BASE_FILENAME" && \
+    echo "Done"
+fi
